net: enetc: force the RGMII speed and duplex when not operating in inband mode
The ENETC port 0 MAC supports in-band status signaling coming from a PHY
when operating in RGMII mode, and this feature is enabled by default.

It has been reported that RGMII is broken in fixed-link, and that is not
surprising considering the fact that no PHY is attached to the MAC in
that case, but a switch.

This brings us to the topic of the patch: the enetc driver should have
not enabled the optional in-band status signaling for RGMII unconditionally.
The feature is still kept, but guarded by the 'managed = "in-band-status"'
DT property. When operating in MLO_AN_PHY and MLO_AN_FIXED modes, we
should not have the in-band signaling enabled, but instead we should
force the speed and duplex to what was resolved by phylink.

Fixes: 71b77a7a27a3 ("enetc: Migrate to PHYLINK and PCS_LYNX")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>